APIMF 2997 js

+++ amf-shapes/shared/src/main/scala/amf/plugins/document/apicontract/parser/spec/common/DataNodeParser.scala
@@ -27,7 +27,7 @@ import scala.collection.mutable.ListBuffer
 trait DataNodeParserContext {
   def findAnnotation(key: String, scope: SearchScope.Scope): Option[CustomDomainProperty]
   def refs: Seq[ParsedReference]
-  def getMaxYamlReferences: Option[Long]
+  def getMaxYamlReferences: Option[Int]
   def fragments: Map[String, FragmentRef]
 }
